Update README.md
Iteration 2:

The user stories that we were able to implement were drawing, color, and background. The drawing function can be found from the menu and can allow you to draw on the canvas. The color function can be accessed from the menu and changes to different colors of the pen and to get out of that activity, you can press the back button. The background function can be accessed from the menu and changes the background of the canvas to a different color and to get out of the activity, you can press the back button. The most difficult task we faced was understanding the unit testing and applying it to the methods in each of our scenarios.